test fixes

+++ tests/cn_vip_testsuite/pointer_copy_user_dataflow_direct_bytewise.pass.c
@@ -7,14 +7,14 @@ int  x=1;
 void user_memcpy(byte *dest, byte *src, size_t n)
 /*@
 requires
-    take Src = each (u64 i; 0u64 <= i && i < n ) { RW(array_shift(src, i)) };
-    take Dest = each (u64 i; 0u64 <= i && i < n ) { W(array_shift(dest, i)) };
+    take Src = each (integer i; 0 <= i && i < n ) { RW(array_shift(src, i)) };
+    take Dest = each (integer i; 0 <= i && i < n ) { W(array_shift(dest, i)) };
 
 ensures
-    take SrcR = each (u64 i; 0u64 <= i && i < n ) { RW(array_shift(src, i)) };
-    take DestR = each (u64 i; 0u64 <= i && i < n ) { RW(array_shift(dest, i)) };
+    take SrcR = each (integer i; 0 <= i && i < n ) { RW(array_shift(src, i)) };
+    take DestR = each (integer i; 0 <= i && i < n ) { RW(array_shift(dest, i)) };
     Src == SrcR;
-    each (u64 i; 0u64 <= i && i < n ) { SrcR[i] == DestR[i] };
+    each (integer i; 0 <= i && i < n ) { SrcR[i] == DestR[i] };
 @*/
 {
   while (n > 0u)
@@ -23,18 +23,19 @@ ensures
     let src_start = {src}@start;
     let dest_start = {dest}@start;
     let n_start = {n}@start;
-    0u64 <= n; n <= n_start;
+    0 <= n; n <= n_start;
     src == array_shift<unsigned char>(src_start, n_start - n);
     dest == array_shift<unsigned char>(dest_start, n_start - n);
-    take S = each (u64 i; 0u64 <= i && i < n_start ) { RW(array_shift(src_start, i)) };
+    take S = each (integer i; 0 <= i && i < n_start ) { RW(array_shift(src_start, i)) };
     Src == S;
-    take D1 = each (u64 i; n_start - n <= i && i < n_start ) { W(array_shift(dest_start, i)) };
-    take D2 = each (u64 i; 0u64 <= i && i < n_start - n ) { RW(array_shift(dest_start, i)) };
-    each (u64 i; 0u64 <= i && i < n_start - n ) { S[i] == D2[i] };
+    take D1 = each (integer i; n_start - n <= i && i < n_start ) { W(array_shift(dest_start, i)) };
+    take D2 = each (integer i; 0 <= i && i < n_start - n ) { RW(array_shift(dest_start, i)) };
+    each (integer i; 0 <= i && i < n_start - n ) { S[i] == D2[i] };
   @*/
   {
     /*@ focus RW<byte>, n_start - n; @*/
     /*@ focus W<byte>, n_start - n; @*/
+    /*@ instantiate (n_start - n); @*/
     *dest = *src;
     src += 1; dest += 1; n -= 1u;
   }
